requirements.txt
line-bot-sdk
flask
gunicorn
psycopg2
lxml
PyMySQL
requests
beautifulsoup4
bs4
pycparser
line-bot-sdk
simplejson
numpy
wordcutpy
pythainlp

